pyproject.toml
[tool.poetry]
name = "late-potentials-research"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [{include = "late_potentials_research"}]

[tool.poetry.dependencies]
python = "^3.11"
numpy = "^1.24.0"
pandas = "^1.5.2"
matplotlib = "^3.6.2"
wfdb = "^4.1.0"
scipy = "^1.9.3"
neurokit2 = "^0.2.2"
pywavelets = "^1.4.1"
colorama = "^0.4.6"
fpdf = "^1.7.2"
python-docx = "^0.8.11"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
